pyproject.toml
[project]
name = "real-time-face-mask-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "gradio>=5.38.2",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "req>=1.0.0",
    "requests>=2.32.4",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "uvicorn>=0.35.0",
]
